Move IntoOwnedContent and AsBorrowedContent
These are currently only used in `filters.rs`, so put them there now.
They can be moved again in the future if they're used more broadly.
